Refactor abstract IT and fix one IT
