+++ pass-deposit-services/deposit-messaging/src/test/java/org/eclipse/pass/deposit/AbstractDepositSubmissionIT.java
@@ -15,12 +15,9 @@
  */
 package org.eclipse.pass.deposit;
 
-import static okhttp3.Credentials.basic;
-import static okhttp3.RequestBody.create;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.util.LinkedList;
 import java.util.List;
@@ -28,11 +25,8 @@ import java.util.Set;
 import java.util.function.BiPredicate;
 import java.util.function.Predicate;
 
-import okhttp3.MediaType;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
 import org.eclipse.deposit.util.async.Condition;
+import org.eclipse.pass.deposit.messaging.DepositApp;
 import org.eclipse.pass.deposit.util.SubmissionTestUtil;
 import org.eclipse.pass.support.client.PassClient;
 import org.eclipse.pass.support.client.model.AggregatedDepositStatus;
@@ -41,88 +35,56 @@ import org.eclipse.pass.support.client.model.PassEntity;
 import org.eclipse.pass.support.client.model.Repository;
 import org.eclipse.pass.support.client.model.Source;
 import org.eclipse.pass.support.client.model.Submission;
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.rules.ExpectedException;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
 
 /**
- * Provides convenience methods for depositing Submission graphs in Fedora.  <strong>N.B.</strong> this test fixture
- * works best coupled with the {@code SpringJunit4ClassRunner}.  If this fixture is <em>not</em> using the Spring
- * Runner, then the caller <em>must</em> set the following properties for proper operation:
- * <ul>
- *     <li>{@link #fcrepoBaseUrl}, {@link #fcrepoUser}, {@link #fcrepoPass}</li>
- *     <li>{@link #contextUri}</li>
- *     <li>{@link #passClient}</li>
- *     <li>{@link #okHttp} always supplied by {@link #setUpOkHttp()} (this client must have an {@code Interceptor}
- *         configured for authenticating to Fedora)</li>
- * </ul>
- * <p>
- * When <em>using</em> the Spring Runner, the dependencies are either {@code @Autowired} or supplied by {@code @Value}.
- * </p>
- * <ul>
- *      <li>{@link #fcrepoBaseUrl}, {@link #fcrepoUser}, {@link #fcrepoPass} supplied by the respective {@code @Value}
- *           expressions {@code ${pass.fedora.baseurl}}, {@code ${pass.fedora.user}},
- *           {@code ${pass.fedora.password}}</li>
- *      <li>{@link #contextUri} supplied by the {@code @Value} expression {@code ${pass.jsonld.context}}</li>
- *      <li>{@link #passClient} supplied by {@code @Autowired}</li>
- *      <li>{@link #okHttp} always supplied by {@link #setUpOkHttp()} (this client must have an {@code Interceptor}
- *          configured for authenticating to Fedora)</li>
- *  </ul>
+ * Provides convenience methods for depositing Submissions in PASS.
  *
  * @author Elliot Metsger (emetsger@jhu.edu)
  */
-public abstract class AbstractSubmissionFixture {
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(classes = DepositApp.class)
+@TestPropertySource(properties = {
+    "pass.client.url=http://localhost:8080",
+    "pass.client.user=backend",
+    "pass.client.password=backend"
+})
+@Testcontainers
+@DirtiesContext
+public abstract class AbstractDepositSubmissionIT {
 
-    @Rule
-    public ExpectedException thrown = ExpectedException.none();
+    private static final DockerImageName PASS_CORE_IMG =
+        DockerImageName.parse("ghcr.io/eclipse-pass/pass-core-main");
 
-    private static Logger LOG = LoggerFactory.getLogger(AbstractSubmissionFixture.class);
+    @Container
+    static final GenericContainer<?> PASS_CORE_CONTAINER = new GenericContainer<>(PASS_CORE_IMG)
+        .withEnv("PASS_CORE_BASE_URL", "http://localhost:8080")
+        .withEnv("PASS_CORE_BACKEND_USER", "backend")
+        .withEnv("PASS_CORE_BACKEND_PASSWORD", "backend")
+        .waitingFor(Wait.forHttp("/data/grant"))
+        .withExposedPorts(8080);
 
-    private static final String MERGE_PATCH = "application/merge-patch+json";
-
-    private static final String SUBMIT_TRUE_PATCH = "" +
-                                                    "{\n" + "   \"@id\": \"%s\",\n" +
-                                                    "   \"@context\": \"%s\",\n" +
-                                                    "   \"submitted\": \"true\",\n" +
-                                                    "   \"submissionStatus\": \"submitted\"\n" +
-                                                    "}";
-
-    private static final long TRAVIS_CONDITION_TIMEOUT_MS = 180 * 1000;
-
-    @Autowired private SubmissionTestUtil submissionTestUtil;
+    @Autowired protected SubmissionTestUtil submissionTestUtil;
 
     @Autowired
     protected PassClient passClient;
 
-    protected OkHttpClient okHttp;
-
-    @Value("${pass.fedora.baseurl}")
-    protected String fcrepoBaseUrl;
-
-    @Value("${pass.fedora.user}")
-    protected String fcrepoUser;
-
-    @Value("${pass.fedora.password}")
-    protected String fcrepoPass;
-
-    @Value("${pass.jsonld.context}")
-    protected String contextUri;
-
-    /**
-     * An OkHttp client used to trigger a submission.
-     * <p>
-     * The {@code SubmissionProcessor} will drop messages from the {@link #passClient} (so it doesn't respond to changes
-     * it makes to resources).  This OkHttpClient does not share the user agent string of the {@link #passClient}, so it
-     * can be used to modify a Submission, and have its messages processed by {@code SubmissionProcessor}.
-     * </p>
-     */
-    @Before
-    public void setUpOkHttp() {
-        okHttp = fcrepoClient(fcrepoUser, fcrepoPass);
+    @DynamicPropertySource
+    static void updateProperties(DynamicPropertyRegistry registry) {
+        registry.add("pass.client.url",
+            () -> "http://localhost:" + PASS_CORE_CONTAINER.getMappedPort(8080));
     }
 
     /**
@@ -138,16 +100,14 @@ public abstract class AbstractSubmissionFixture {
      *         Submission.AggregatedDepositStatus.NOT_STARTED}</li>
      * </ul>
      */
-    public List<PassEntity> createSubmission(InputStream submissionGraph) throws IOException {
+    public List<PassEntity> createSubmission(InputStream submissionGraph) {
         List<PassEntity> entities = new LinkedList<>();
 
         // Upload sample data to Fedora repository to get its Submission URI.
         String submissionId = submissionTestUtil.readSubmissionJsonAndAddToPass(submissionGraph, entities).getId();
 
         // Find the Submission entity that was uploaded
-        // TODO Deposit service port pending
-//        Submission submission = findSubmission(uriMap);
-        Submission submission = null;
+        Submission submission = findSubmission(entities);
 
         // verify state of the initial Submission
         assertEquals("Submission must have a Submission.source = Submission.Source.PASS",
@@ -157,10 +117,15 @@ public abstract class AbstractSubmissionFixture {
                      AggregatedDepositStatus.NOT_STARTED, submission.getAggregatedDepositStatus());
 
         // no Deposits pointing to the Submission
-        assertTrue("Unexpected incoming links to " + submissionId,
-                   SubmissionTestUtil.getDepositUris(submission, passClient).isEmpty());
+        // TODO Deposit service port pending
+//        assertTrue("Unexpected incoming links to " + submissionId,
+//                   SubmissionTestUtil.getDepositUris(submission, passClient).isEmpty());
 
         return entities;
+    }
+
+    public void triggerSubmission(String submissionId) {
+
     }
 
     /**
@@ -190,44 +155,6 @@ public abstract class AbstractSubmissionFixture {
             .filter(submissionFilter)
             .findAny()
             .get();
-    }
-
-    /**
-     * Flips the {@code Submission.submitted} flag to {@code true}, at which point Deposit Services will process
-     * the {@code Submission}.  The {@code Submission} should be considered read-only to clients external to the Deposit
-     * Services runtime after invoking this method.   This means <strong>prior</strong> to invoking this method, the
-     * {@code Submission} should be fully linked to all necessary resources, including {@link
-     * org.eclipse.pass.support.client.model.File file} content and any {@link Repository repositories}.  The {@link
-     * Submission#getSource() source} and {@link Submission#getAggregatedDepositStatus() deposit status} should also be
-     * set properly.  If the Ember UI is being emulated then the proper values are:
-     * <dl>
-     *     <dt>source</dt>
-     *     <dd>{@link Source#PASS}</dd>
-     *     <dt>deposit status</dt>
-     *     <dd>{@link AggregatedDepositStatus#NOT_STARTED}</dd>
-     * </dl>
-     * Note: Deposit Services doesn't care about, or examine,
-     * {@link org.eclipse.pass.support.client.model.SubmissionStatus}.
-     *
-     * @param submissionId the ID of a {@code Submission} that is ready for processing
-     */
-    public void triggerSubmission(String submissionId) {
-        String body = String.format(SUBMIT_TRUE_PATCH, submissionId, contextUri);
-
-        Request post = new Request.Builder()
-            .addHeader("Content-Type", MERGE_PATCH)
-            .method("PATCH", create(MediaType.parse(MERGE_PATCH), body))
-            .url(submissionId)
-            .build();
-
-        try (Response response = okHttp.newCall(post).execute()) {
-            int expected = 204;
-            assertEquals("Triggering 'submitted' flag to 'true' for " + submissionId + " failed.  " +
-                         "Expected " + expected + ", got " + response.code() + " (" + response.message() + ")",
-                         expected, response.code());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
     }
 
     /**
@@ -319,52 +246,4 @@ public abstract class AbstractSubmissionFixture {
 //                                 "Missing Deposit for Repository " + repositoryUri + " on Submission " +
 //                                 submission.getId()));
 //    }
-
-    /**
-     * Method for determining whether the runtime platform is Travis.  Useful for conditionally extending timeouts.
-     *
-     * @return true if the runtime platform is the Travis integration environment
-     */
-    public static boolean travis() {
-        if (System.getenv("TRAVIS") != null &&
-            System.getenv("TRAVIS").equalsIgnoreCase("true")) {
-            return true;
-        }
-
-        if (System.getProperty("TRAVIS") != null && System.getProperty("TRAVIS").equalsIgnoreCase("true")) {
-            return true;
-        }
-
-        if (System.getProperty("travis") != null && System.getProperty("travis").equalsIgnoreCase("true")) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Answers an OkHttpClient for communicating with a Fedora repository.  It will automatically supply an {@code
-     * Authorization} header on the HTTP response if required.
-     *
-     * @param fcrepoUser the user to use for authentication to fedora, may be {@code null} if Fedora doesn't require
-     *                   authn
-     * @param fcrepoPass the password to use for authentication to fedora, may be {@code null} if Fedora doesn't require
-     *                   authn
-     * @return an OkHttpClient configured for communicating with a Fedora repository
-     */
-    public OkHttpClient fcrepoClient(String fcrepoUser, String fcrepoPass) {
-        OkHttpClient.Builder builder = new OkHttpClient.Builder();
-        builder.authenticator((route, response) -> {
-            if (response.header("Authorization") != null) {
-                return null;
-            }
-            return response
-                .request()
-                .newBuilder()
-                .header("Authorization", basic(fcrepoUser, fcrepoPass))
-                .build();
-        });
-
-        return builder.build();
-    }
 }
